@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {AuthService} from "./services/auth.service";
-import {Observable} from "rxjs/Observable";
+import {Observable} from "rxjs";
 import {User} from "./model/user";
 
 @Component({
