@@ -1,8 +1,9 @@
+
+import {shareReplay, filter, tap, map} from 'rxjs/operators';
 import {Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {Observable} from "rxjs/Observable";
+import {Observable, BehaviorSubject} from "rxjs";
 import {User} from "../model/user";
-import {BehaviorSubject} from "rxjs/BehaviorSubject";
 
 export const ANONYMOUS_USER: User = {
     id: undefined,
@@ -16,41 +17,41 @@ export class AuthService {
 
     private subject = new BehaviorSubject<User>(undefined);
 
-    user$: Observable<User> = this.subject.asObservable().filter(user => !!user);
+    user$: Observable<User> = this.subject.asObservable().pipe(filter(user => !!user));
 
-    isLoggedIn$: Observable<boolean> = this.user$.map(user => !!user.id);
+    isLoggedIn$: Observable<boolean> = this.user$.pipe(map(user => !!user.id));
 
-    isLoggedOut$: Observable<boolean> = this.isLoggedIn$.map(isLoggedIn => !isLoggedIn);
+    isLoggedOut$: Observable<boolean> = this.isLoggedIn$.pipe(map(isLoggedIn => !isLoggedIn));
 
     constructor(private http: HttpClient) {
-        http.get<User>('/api/user')
-            .do(console.log)
+        http.get<User>('/api/user').pipe(
+            tap(console.log))
             .subscribe(user => this.subject.next(user ? user : ANONYMOUS_USER));
     }
 
     signUp(email:string, password:string ) {
 
-        return this.http.post<User>('/api/signup', {email, password})
-            .shareReplay()
-            .do(user => this.subject.next(user));
+        return this.http.post<User>('/api/signup', {email, password}).pipe(
+            shareReplay(),
+            tap(user => this.subject.next(user)),);
     }
 
     login(email:string, password:string ) {
-        return this.http.post<User>('/api/login', {email, password})
-            .shareReplay()
-            .do(user => this.subject.next(user));
+        return this.http.post<User>('/api/login', {email, password}).pipe(
+            shareReplay(),
+            tap(user => this.subject.next(user)),);
     }
 
     loginAsUser(email:string) {
-        return this.http.post<User>('/api/admin', {email})
-            .shareReplay()
-            .do(user => this.subject.next(user));
+        return this.http.post<User>('/api/admin', {email}).pipe(
+            shareReplay(),
+            tap(user => this.subject.next(user)),);
     }
 
     logout() : Observable<any> {
-        return this.http.post('/api/logout', null)
-            .shareReplay()
-            .do(user => this.subject.next(ANONYMOUS_USER));
+        return this.http.post('/api/logout', null).pipe(
+            shareReplay(),
+            tap(user => this.subject.next(ANONYMOUS_USER)),);
     }
 }
 
