@@ -1,6 +1,8 @@
 
+import {tap, first, map} from 'rxjs/operators';
+
 import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from "@angular/router";
-import {Observable} from "rxjs/Observable";
+import {Observable} from "rxjs";
 import {AuthService} from "./auth.service";
 import * as _ from 'lodash';
 import {Injectable} from "@angular/core";
@@ -19,14 +21,14 @@ export class AuthorizationGuard implements  CanActivate {
         route: ActivatedRouteSnapshot,
         state: RouterStateSnapshot): Observable<boolean>  {
 
-        return this.authService.user$
-            .map(user => _.intersection(this.allowedRoles, user.roles).length > 0 )
-            .first()
-            .do(allowed => {
+        return this.authService.user$.pipe(
+            map(user => _.intersection(this.allowedRoles, user.roles).length > 0 ),
+            first(),
+            tap(allowed => {
                 if (!allowed) {
                     this.router.navigateByUrl('/');
                 }
-            });
+            }),);
 
 
     }
