@@ -1,8 +1,10 @@
 
+import {map} from 'rxjs/operators';
+
 import {Injectable} from "@angular/core";
 import {HttpClient} from "@angular/common/http";
 import {Lesson} from "../model/lesson";
-import {Observable} from "rxjs/Observable";
+import {Observable} from "rxjs";
 
 
 @Injectable()
@@ -13,8 +15,8 @@ export class LessonsService {
     }
 
     loadAllLessons() : Observable<Lesson[]> {
-        return this.http.get<any>('/api/lessons')
-            .map(res => res.lessons);
+        return this.http.get<any>('/api/lessons').pipe(
+            map(res => res.lessons));
     }
 
     findLessonById(id:number) {
