@@ -15,14 +15,14 @@ import {AdminComponent} from './admin/admin.component';
 import {Router, RouterModule} from "@angular/router";
 
 
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/shareReplay';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/filter';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/first';
-import 'rxjs/add/observable/of';
+
+
+
+
+
+
+
+
 import {RbacAllowDirective} from "./common/rbac-allow.directive";
 import {AuthorizationGuard} from "./services/authorization.guard";
 
